Set apex_available property
The marked library(ies) were available to the APEXes via the
hand-written whitelist in build/soong/apex/apex.go. Trying to remove the
whitelist by adding apex_available property to the Android.bp of the
libraries.

This is actually the cherry-pick of ag/11305332 which was uploaded
when libziparchive was in /system/core/libziparchive. It looks like that
the change was missed while the library was moved to
/system/libziparchive.

Exempt-From-Owner-Approval: approved by baligh@

Bug: 150999716
Test: m
Change-Id: I67c667a5041bb0743a4f0c486287b16667194717