Disable tests that depend on iteration order.
Because libziparchive implements iteration by going through its hash,
iteration order isn't the same as the order in which entries were found
in the file. That's why these tests are so sparse on tests that involve
multiple zip entries. These two snuck in because on LP64 our hash
iteration order just happens to match file iteration order, but the LP32
hash differs and the tests bogusly fail there.

Disable them for now, because there's no easy fix for this, especially
because we don't want to have to pay to preserve iteration order for the
normal use case of this library: pulling stuff out of apks by name (with
no iteration at all).

Test: cli-test unzip.test
Change-Id: Icb30462eb80346a367e87b46936c5d19db2a10f7