[incfs] SIGBUS handling in libziparchive
This change adds the protections required to read APKs residing
on IncFS filesystem. Moved libhardening code into libziparchive
and created a new library, libziparchive_for_incfs, to handle
the potential SIGBUS errors when hitting a not-yet-loaded page
in a mmap-ped file.
This new library also exports the primitves needed for a general
hardening

Bug: 160635079
Test: atest ziparchive-tests
Change-Id: I7a8742317d66c4a505c9cd407540ee1ce3b46ac7